+++ vanilla-js/public/images/favicon.png
