--- vanilla-js/vApp.css
@@ -1,159 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Manrope:wght@300;800&display=swap");
-
-.App {
-  text-align: center;
-}
-body {
-  background-color: #000000;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 94vh;
-  font-family: "Manrope", sans-serif;
-}
-
-.card {
-  background-color: hsl(56, 100%, 56%);
-  margin: auto;
-  width: 550px;
-  height: 300px;
-  border-radius: 0 50px 50px 50px;
-  padding: 0 1rem 1rem 1rem;
-  box-shadow: inset 0px 0px 15px hsl(9, 93%, 53%);
-  box-shadow: inset 0px 0px 5px hsl(9, 98%, 34%);
-}
-.card-content {
-  margin: 0;
-  position: relative;
-  top: 50%;
-  -ms-transform: translateY(-50%);
-  transform: translateY(-50%);
-}
-
-.icon {
-  width: 50px;
-}
-
-div.img-area {
-  padding-top: 1.5rem;
-  height: 50px;
-}
-
-div.advice-area {
-  min-height: 270px;
-}
-#advice {
-  font-size: 1.4em;
-  color: hsl(0, 0%, 0%);
-  margin-bottom: 2rem;
-  margin-top: 1rem;
-  font-family: "Manrope", sans-serif;
-  font-weight: 800;
-}
-
-#advice-type {
-  color: hsl(9, 100%, 42%);
-  text-transform: uppercase;
-  font-family: "Manrope", sans-serif;
-  font-weight: 300;
-}
-
-div.link {
-  font-size: 1rem;
-  padding-bottom: 1rem;
-}
-
-select {
-  background-color: hsl(46, 100%, 56%);
-  border: none;
-  padding: 0.3rem;
-  width: 130px;
-}
-
-input[type="checkbox"] {
-  display: none;
-}
-.check-icon {
-  display: inline-block;
-  width: 15px;
-  height: 15px;
-  background-image: url(./images/checkbox--unchecked.svg);
-  background-repeat: no-repeat;
-  background-size: 15px;
-}
-
-input[type="checkbox"]:checked + .check-icon {
-  background-image: url(./images/checkbox--checked.svg);
-}
-
-#remove-boredom-button {
-  width: auto;
-  height: 30px;
-  padding: 0 1rem;
-  font-weight: 800;
-  letter-spacing: 1px;
-  border-radius: 0 20px 20px 20px;
-  background-color: hsl(9, 100%, 42%);
-  color: hsl(56, 100%, 56%);
-  border: none;
-  position: relative;
-  top: -15px;
-  transition: 0.3s;
-}
-
-#remove-boredom-button:active {
-  background-color: rgb(45, 145, 95);
-  box-shadow: 0px 0px 20px rgb(45, 145, 95) !important;
-}
-
-#remove-boredom-button:hover {
-  box-shadow: 0px 0px 20px hsl(9, 100%, 42%);
-}
-form {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  gap: 0.7rem;
-}
-
-#credits {
-  color: #4e4e4e;
-  text-align: center;
-  position: relative;
-  bottom: 0;
-  padding-bottom: 0rem;
-  display: flex;
-  justify-content: center;
-}
-
-#credits:hover .dropdown-content {
-  display: block;
-}
-
-.dropdown-content {
-  display: none;
-  position: absolute;
-  background-color: #f9f9f9;
-  min-width: 160px;
-  padding: 12px 16px;
-  z-index: 1;
-  bottom: 1rem;
-  border-radius: 20px;
-}
-.dropdown-content ul {
-  text-align: left;
-  font-size: 0.8rem;
-}
-
-@media only screen and (max-width: 975px) {
-  .card {
-    width: 300px;
-    height: 360px;
-  }
-  form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 0.2rem;
-  }
-}
