+++ vanilla-js/webpack.config.js
@@ -1,0 +1,43 @@
+/*module.exports = {
+  mode: "development",
+  devServer: {
+    contentBase: "./dist",
+  },
+  module: {
+    rules: [
+      {
+        test: /\.js$/,
+        use: [{ loader: "babel-loader" }],
+        exclude: /node_modules/,
+      },
+    ],
+  },
+};
+
+*/
+const webpack = require("webpack");
+const path = require("path");
+
+const config = {
+  mode: "development",
+  entry: "./src/index.js",
+  output: {
+    path: path.resolve(__dirname, "dist"),
+    filename: "bundle.js",
+  },
+  module: {
+    rules: [
+      {
+        test: /\.js$/,
+        use: "babel-loader",
+        exclude: /node_modules/,
+      },
+      {
+        test: /\.css$/,
+        use: ["style-loader", "css-loader"],
+      },
+    ],
+  },
+};
+
+module.exports = config;
